+++ modelgym/rf_model.py
@@ -35,12 +35,11 @@ class RFModel(Model):
         return ab
 
     def fit(self, params, dtrain, dtest, n_estimators):
-        rf = rfc(n_estimators=params['n_estimators'], max_depth=params['max_depth'], criterion=params['criterion'],
-                 verbose=params['verbose'])
-        bst = rf.fit(dtrain.X, dtrain.y)
-        res = rf.predict(dtest.X)
-        print("\n", res)
-        return bst, res
+        rf = rfc(random_state=0, n_jobs=-1, n_estimators=params['n_estimators'], max_depth=params['max_depth'],
+                 criterion=params['criterion']).fit(dtrain.X, dtrain.y)
+        preds = rf.predict(dtest.X)
+        # print("\n", res)
+        return rf, preds
 
     def predict(self, bst, dtest, X_test):
         preds = bst.predict(dtest.X)
